[#61] fixes after rebase
